+++ framework/rbac/DbManager.php
@@ -87,6 +87,11 @@ class DbManager extends BaseManager
      * @since 2.0.3
      */
     public $cacheKey = 'rbac';
+    /**
+     * @var string the key used to store user RBAC roles in cache
+     * @since 2.0.48
+     */
+    public $rolesCacheSuffix = 'roles';
 
     /**
      * @var Item[] all auth items (name => Item)
@@ -471,6 +476,14 @@ class DbManager extends BaseManager
             return [];
         }
 
+        if ($this->cache !== null) {
+            $data = $this->cache->get($this->getUserRolesCacheKey());
+
+            if ($data !== false) {
+                return $data;
+            }
+        }
+
         $query = (new Query())->select('b.*')
             ->from(['a' => $this->assignmentTable, 'b' => $this->itemTable])
             ->where('{{a}}.[[item_name]]={{b}}.[[name]]')
@@ -480,6 +493,10 @@ class DbManager extends BaseManager
         $roles = $this->getDefaultRoleInstances();
         foreach ($query->all($this->db) as $row) {
             $roles[$row['name']] = $this->populateItem($row);
+        }
+
+        if ($this->cache !== null) {
+            $this->cache->set($this->getUserRolesCacheKey(), $roles);
         }
 
         return $roles;
@@ -985,6 +1002,7 @@ class DbManager extends BaseManager
     {
         if ($this->cache !== null) {
             $this->cache->delete($this->cacheKey);
+            $this->cache->delete($this->getUserRolesCacheKey());
             $this->items = null;
             $this->rules = null;
             $this->parents = null;
@@ -1060,5 +1078,10 @@ class DbManager extends BaseManager
     protected function isEmptyUserId($userId)
     {
         return !isset($userId) || $userId === '';
+    }
+
+    private function getUserRolesCacheKey()
+    {
+        return $this->cacheKey . $this->rolesCacheSuffix;
     }
 }